requirements.txt
setuptools
pydantic
azure-storage-blob
azure-core
python-dotenv
openai
pytest